* [core]: better check before claim colour managed
The transform is checked additional to the previously ICC profile before
setting the _NET_COLOR_DESKTOP atom to signal conversions are in effect.
This catches cases, where Oyranos can not create or perform a colour
transform. (Thats the case with actual Oyranos git.)

This fixes the qcmsevents icon showing a coloured icon, while it should show
a gray one for none working colour server.