* [core]: support per region profiles
... alias _NET_COLOR_PROFILES as originally intented by Tomas Carnecky.
The original implementation is improved in many ways, such as working multiple
profiles per server, multiple regions per window and device link support.

It is regressed regarding the _NET_COLOR_TARGET atom, which might be useful.

_NET_COLOR_MANAGEMENT is not supported and is not likely to be.
This was a requirement during GSoC, but after some experience it seems a region
without any window profile has the same effect of switching colour management
off for a region.