* [core]: compute stencil ID during drawing
Omit the global stencil ID assignment per window and region.

The new colour_desktop_region_count variable signals with -1 to need
initialisation. forEachWindowOnScreen() crashes inside the event loop in
pluginHandleEvent(), so it was moved inside pluginDrawWindow().

PrivateWindw::stencil_id is removed. Is is not used for testing for
window regions. HAS_REGIONS() is added for the tests.

Some debug warnings are added in updateWindowRegions().

The stencil pool was not a good idea, and code looks much better without.
The stencil IDs are assigned per region and per monitor. This means, the
X Color Management spec can not copy completely with that situation. This is
normally not mentionable, but device link profiles can not completely be
adressed. So users can currently set a custom transform only for one monitor
inside multi monitor systems in the CompICC colour server.